dnlkwnr

+++ send_data/views.py
@@ -91,6 +91,8 @@ class TokenDestroyView(APIView):
     def post(self, request):
         try:
             token = Token.objects.get(key=self.request.META.get('HTTP_AUTHORIZATION', None))
+            if token == '' or token == NULL or token == ' ':
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
             token.delete()
         except:            
             return Response({"detail":"Authentication credentials were not provided."}, status=status.HTTP_401_UNAUTHORIZED)
